requirements.txt
bpython
pdbpp
cryptography
qrcode-terminal
docopt
